// frontend/sdk/models/FameTier.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sopranino API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const FameTier = {
    Easy: 'EASY',
    Medium: 'MEDIUM',
    Hard: 'HARD',
    Expert: 'EXPERT',
    Impossible: 'IMPOSSIBLE'
} as const;
export type FameTier = typeof FameTier[keyof typeof FameTier];


export function instanceOfFameTier(value: any): boolean {
    for (const key in FameTier) {
        if (Object.prototype.hasOwnProperty.call(FameTier, key)) {
            if (FameTier[key as keyof typeof FameTier] === value) {
                return true;
            }
        }
    }
    return false;
}

export function FameTierFromJSON(json: any): FameTier {
    return FameTierFromJSONTyped(json, false);
}

export function FameTierFromJSONTyped(json: any, ignoreDiscriminator: boolean): FameTier {
    return json as FameTier;
}

export function FameTierToJSON(value?: FameTier | null): any {
    return value as any;
}

export function FameTierToJSONTyped(value: any, ignoreDiscriminator: boolean): FameTier {
    return value as FameTier;
}